# backend/docs/openapi.yaml
openapi: 3.1.0

info:
  title: Aether API
  version: 0.1.0
  description: |
    HTTP API for Aether, a modular monolith backend for a real-time
    communication system.

servers:
  - url: http://localhost:8080
    description: Local development

tags:
  - name: Authentication
    description: User registration and authentication operations.
  - name: Users
    description: User lookup and profile operations.
  - name: Contacts
    description: Contact request and relationship operations.

paths:
  /register:
    post:
      operationId: registerUser
      tags:
        - Authentication
      summary: Register a user
      description: Creates a user after normalizing and validating the supplied registration data.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
            examples:
              default:
                summary: Registration request
                value:
                  email: user@example.com
                  username: petar
                  password: password123
                  firstName: Petar
                  lastName: Petrović
      responses:
        "201":
          description: User registered successfully.
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /login:
    post:
      operationId: loginUser
      tags:
        - Authentication
      summary: Log in a user
      description: |
        Verifies an email and password and returns a short-lived JWT access token.
        The email is trimmed and lowercased; password whitespace is preserved.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            examples:
              default:
                summary: Login request
                value:
                  email: user@example.com
                  password: password123
      responses:
        "200":
          description: Login successful.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
              examples:
                default:
                  summary: Access token response
                  value:
                    accessToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.example.signature
                    tokenType: Bearer
                    expiresIn: 900
        "400":
          description: The request body is malformed.
          content:
            text/plain:
              schema:
                type: string
              example: "invalid request body\n"
        "401":
          $ref: "#/components/responses/InvalidCredentials"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /users:
    get:
      operationId: getUserByEmail
      tags:
        - Users
      summary: Get a user by email
      parameters:
        - name: email
          in: query
          required: true
          description: Email address of the user to retrieve.
          schema:
            type: string
            format: email
          example: user@example.com
      responses:
        "200":
          description: User found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
              examples:
                default:
                  summary: User response
                  value:
                    id: 550e8400-e29b-41d4-a716-446655440000
                    email: user@example.com
                    username: petar
                    firstName: Petar
                    lastName: Petrović
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          description: User not found.
          content:
            text/plain:
              schema:
                type: string
              example: "user not found\n"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /contact-requests:
    post:
      operationId: sendContactRequest
      tags:
        - Contacts
      summary: Send a contact request
      description: Sends a contact request to the user with the supplied username.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendContactRequest"
            example:
              username: petar
      responses:
        "201":
          description: Contact request sent.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactRequestCreated"
              example:
                id: 550e8400-e29b-41d4-a716-446655440001
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: No user has the supplied username.
        "409":
          description: A pending request already exists between the two users.
        "500":
          $ref: "#/components/responses/InternalServerError"

  /contact-requests/{requestID}:
    delete:
      operationId: cancelContactRequest
      tags:
        - Contacts
      summary: Cancel a contact request
      description: Cancels a pending request sent by the authenticated user.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ContactRequestID"
      responses:
        "204":
          description: Contact request cancelled.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: A matching pending contact request was not found.
        "500":
          $ref: "#/components/responses/InternalServerError"

  /contact-requests/{requestID}/accept:
    post:
      operationId: acceptContactRequest
      tags:
        - Contacts
      summary: Accept a contact request
      description: Accepts a pending request received by the authenticated user and creates the contact relationship.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ContactRequestID"
      responses:
        "204":
          description: Contact request accepted.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: A matching pending contact request was not found.
        "500":
          $ref: "#/components/responses/InternalServerError"

  /contact-requests/{requestID}/decline:
    post:
      operationId: declineContactRequest
      tags:
        - Contacts
      summary: Decline a contact request
      description: Declines a pending request received by the authenticated user.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ContactRequestID"
      responses:
        "204":
          description: Contact request declined.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: A matching pending contact request was not found.
        "500":
          $ref: "#/components/responses/InternalServerError"

components:
  parameters:
    ContactRequestID:
      name: requestID
      in: path
      required: true
      description: Contact request identifier.
      schema:
        type: string
        format: uuid
      example: 550e8400-e29b-41d4-a716-446655440001

  schemas:
    SendContactRequest:
      type: object
      additionalProperties: false
      required:
        - username
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 30
          description: Username of the request recipient. Surrounding whitespace is removed.
          example: petar

    ContactRequestCreated:
      type: object
      additionalProperties: false
      required:
        - id
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440001

    RegisterRequest:
      type: object
      required:
        - email
        - username
        - password
        - firstName
        - lastName
      properties:
        email:
          type: string
          format: email
          description: Email address, trimmed and lowercased during registration.
          example: user@example.com
        username:
          type: string
          minLength: 3
          maxLength: 30
          description: Unique username after surrounding whitespace is removed.
          example: petar
        password:
          type: string
          format: password
          writeOnly: true
          description: Password containing between 8 and 72 UTF-8 bytes. Whitespace is preserved.
          x-aether-minLengthBytes: 8
          x-aether-maxLengthBytes: 72
          example: password123
        firstName:
          type: string
          minLength: 2
          maxLength: 50
          description: First name after surrounding whitespace is removed.
          example: Petar
        lastName:
          type: string
          minLength: 2
          maxLength: 50
          description: Last name after surrounding whitespace is removed.
          example: Petrović

    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: Email address, trimmed and lowercased before credential lookup.
          example: user@example.com
        password:
          type: string
          format: password
          writeOnly: true
          description: Password whose whitespace is preserved during verification.
          example: password123

    LoginResponse:
      type: object
      additionalProperties: false
      required:
        - accessToken
        - tokenType
        - expiresIn
      properties:
        accessToken:
          type: string
          description: Signed JWT access token.
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.example.signature
        tokenType:
          type: string
          enum:
            - Bearer
          example: Bearer
        expiresIn:
          type: integer
          format: int64
          minimum: 1
          description: Number of seconds until the access token expires.
          example: 900

    User:
      type: object
      additionalProperties: false
      required:
        - id
        - email
        - username
        - firstName
        - lastName
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        email:
          type: string
          format: email
          example: user@example.com
        username:
          type: string
          minLength: 3
          maxLength: 30
          example: petar
        firstName:
          type: string
          example: Petar
        lastName:
          type: string
          example: Petrović

  responses:
    BadRequest:
      description: The request is malformed or contains invalid data.
      content:
        text/plain:
          schema:
            type: string
          examples:
            invalidBody:
              summary: Malformed request body
              value: "invalid request body\n"
            invalidEmail:
              summary: Invalid email address
              value: "invalid email format\n"
            emptyEmail:
              summary: Missing email query parameter
              value: "email is empty\n"

    InternalServerError:
      description: An unexpected internal error occurred.
      content:
        text/plain:
          schema:
            type: string
          example: "internal error\n"

    Unauthorized:
      description: A valid bearer access token is required.
      content:
        text/plain:
          schema:
            type: string
          example: "unauthorized\n"

    InvalidCredentials:
      description: The email or password is incorrect.
      content:
        text/plain:
          schema:
            type: string
          example: "invalid credentials\n"

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supply the access token returned by `/login`.