# backend/internal/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package db

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type Contact struct {
	User1ID   pgtype.UUID
	User2ID   pgtype.UUID
	CreatedAt pgtype.Timestamptz
}

type ContactRequest struct {
	ID          pgtype.UUID
	SenderID    pgtype.UUID
	RecipientID pgtype.UUID
	Status      string
	CreatedAt   pgtype.Timestamptz
	UpdatedAt   pgtype.Timestamptz
}

type Conversation struct {
	ID                  pgtype.UUID
	Name                pgtype.Text
	CreatedBy           pgtype.UUID
	LastMessageSequence int64
	CreatedAt           pgtype.Timestamptz
	UpdatedAt           pgtype.Timestamptz
}

type ConversationParticipant struct {
	ConversationID pgtype.UUID
	UserID         pgtype.UUID
	CreatedAt      pgtype.Timestamptz
	UpdatedAt      pgtype.Timestamptz
}

type Message struct {
	ID              pgtype.UUID
	ConversationID  pgtype.UUID
	SenderID        pgtype.UUID
	ClientMessageID pgtype.UUID
	MessageSequence int64
	Body            string
	CreatedAt       pgtype.Timestamptz
	UpdatedAt       pgtype.Timestamptz
}

type User struct {
	ID           pgtype.UUID
	Username     string
	Email        string
	PasswordHash string
	FirstName    string
	LastName     string
	CreatedAt    pgtype.Timestamptz
	UpdatedAt    pgtype.Timestamptz
}